move accessors to core
