+++ package/MDAnalysis/converters/__init__.py
@@ -31,7 +31,6 @@ trajectory objects from other packages to an MDAnalysis object, and vice-versa.
 .. versionadded:: 2.0.0
 
 """
-from . import accessors
 from . import ParmEdParser
 from . import ParmEd
 from . import RDKitParser
