+++ package/doc/sphinx/source/documentation_pages/converters.rst
@@ -44,5 +44,5 @@ Another syntax is also available for tab-completion support::
 .. toctree::
     :maxdepth: 1
 
-    converters/Accessors
+    core/accessors
