--- package/doc/sphinx/source/documentation_pages/converters/Accessors.rst
@@ -1,1 +1,0 @@
-.. automodule:: MDAnalysis.converters.accessors
