+++ package/doc/sphinx/source/documentation_pages/core/accessors.rst
@@ -1,0 +1,1 @@
+.. automodule:: MDAnalysis.core.accessors
